--- swiftbook-olimp-project/FarmViewController.swift
@@ -1,27 +1,0 @@
-//
-//  FarmViewController.swift
-//  swiftbook-olimp-project
-//
-//  Created by Gleb Kalachev on 11/10/17.
-//  Copyright © 2017 Gleb Kalachev. All rights reserved.
-//
-
-import UIKit
-
-class FarmViewController: UIViewController {
-   
-   @IBOutlet weak var swiftcoinsCountLabel: UILabel!
-   
-   @IBOutlet weak var energyCountLabel: UILabel!
-   
-   @IBOutlet weak var minerImageView: UIImageView!
-   
-   @IBOutlet weak var farmButton: UIButton!
-   
-   
-   
-   
-   @IBAction func farmButtonTapped(_ sender: UIButton) {
-   }
-   
-}
